Reduce scalar_traits.h
