+++ src/mpi/mpi_datatype.h
@@ -13,8 +13,6 @@
 #ifndef QMCPLUSPLUS_MPI_DATATYPEDEFINE_H
 #define QMCPLUSPLUS_MPI_DATATYPEDEFINE_H
 
-
-#include "type_traits/scalar_traits.h"
 #if defined(HAVE_MPI)
 #include <mpi.h>
 #else
